user_characters_db.go を作成する
